fix some errors

+++ backend/src/main/java/org/project/backend/service/PatientService.java
@@ -6,25 +6,35 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class PatientService {
+
     @Autowired
     private PatientRepository patientRepository;
 
     public List<Patient> getAllPatients() {
         return patientRepository.findAll();
     }
+
     public Patient addPatient(Patient patient) {
         return patientRepository.save(patient);
     }
-    public Patient getPatientById(Long id) {
-        return patientRepository.getPatientById(id);
+
+    public Optional<Patient> getPatientById(Long id) {
+        return patientRepository.findById(id);
     }
+
     public Patient updatePatient(Patient patient) {
         return patientRepository.save(patient);
     }
-    public void deletePatient(Long id) {
-        patientRepository.deleteById(id);
+
+    public boolean deletePatient(Long id) {
+        if (patientRepository.existsById(id)) {
+            patientRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 }
